last changes for deploy

+++ frontend-quitanda/src/services/UserService.jsx
@@ -1,13 +1,15 @@
-import axios from 'axios';
-import { handleError } from '../utils/handleErrors';
-import { LocalStorageKeys } from '../enums/local-storage-keys-enum';
+import axios from "axios";
+import { handleError } from "../utils/handleErrors";
+import { LocalStorageKeys } from "../enums/local-storage-keys-enum";
 
-const url = 'http://localhost:8000';
+const url = "https://api-quitanda.herokuapp.com";
 
 const verifyToken = () => {
   const localUser = JSON.parse(localStorage.getItem(LocalStorageKeys.USER));
   if (localUser) {
-    axios.defaults.headers.common['Authorization'] = `Bearer ${localUser.token}`;
+    axios.defaults.headers.common[
+      "Authorization"
+    ] = `Bearer ${localUser.token}`;
     return true;
   } else {
     return false;
@@ -38,42 +40,41 @@ export async function addUser(user) {
     return await axios.post(`${url}/users`, user);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function getAllValidUsersWithPagination(sortConfig, page) {
   try {
     return await axios.get(`${url}/valid_users/${sortConfig}/${page}`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function getAllValidUsers() {
   try {
     return await axios.get(`${url}/valid_users`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function getUserByUsername(username) {
   try {
     return await axios.get(`${url}/user/${username}`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 // AUTH ROUTES
 export async function getAllUsers() {
   try {
-    if (verifyToken())
-      return await axios.get(`${url}/users`);
+    if (verifyToken()) return await axios.get(`${url}/users`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function getAllUsersWithPagination(sortConfig, page) {
   try {
@@ -81,17 +82,16 @@ export async function getAllUsersWithPagination(sortConfig, page) {
       return await axios.get(`${url}/users/${sortConfig}/${page}`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function getAllInvalidUsers() {
   try {
-    if (verifyToken())
-      return await axios.get(`${url}/invalid_users`);
+    if (verifyToken()) return await axios.get(`${url}/invalid_users`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function getAllInvalidUsersWithPagination(sortConfig, page) {
   try {
@@ -99,8 +99,8 @@ export async function getAllInvalidUsersWithPagination(sortConfig, page) {
       return await axios.get(`${url}/invalid_users/${sortConfig}/${page}`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function updateUserByUsername(username, userData) {
   try {
@@ -108,23 +108,27 @@ export async function updateUserByUsername(username, userData) {
       return await axios.patch(`${url}/user/${username}`, userData);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
 export async function deleteUserByUsername(username) {
   try {
-    if (verifyToken())
-      return await axios.delete(`${url}/users/${username}`);
+    if (verifyToken()) return await axios.delete(`${url}/users/${username}`);
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
 
-export async function updateAccessAuthorization(username, access_authorization) {
+export async function updateAccessAuthorization(
+  username,
+  access_authorization
+) {
   try {
     if (verifyToken())
-      return await axios.patch(`${url}/user/${username}/access_authorization`, { "access_authorization": access_authorization });
+      return await axios.patch(`${url}/user/${username}/access_authorization`, {
+        access_authorization: access_authorization,
+      });
   } catch (error) {
     handleError(error);
-  };
-};
+  }
+}
